requirements.txt
# Core Dependencies
streamlit==1.52.1
pandas==2.3.3
numpy==2.2.6
plotly==6.5.0

# Trading & Market Data
ccxt==4.5.28
python-binance==1.0.22
ta==0.11.0

# Async & HTTP
aiohttp==3.13.2
requests==2.32.5
nest-asyncio==1.6.0

# Configuration
python-dotenv==1.2.1
PyYAML==6.0.2

# System & Process Management
psutil==7.1.3

# Telegram Integration
python-telegram-bot==21.10

# Payment Processing
stripe==11.2.0
flask==3.0.0

# Utilities
python-dateutil==2.9.0
